requirements.txt
functions-framework>=3.0.0
google-genai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
flask>=2.0.0
